change data type for ParSortID

+++ src/Feedback/FB_Init.cpp
@@ -12,7 +12,7 @@ void FB_Init_SNe();
 void (*FB_Init_User_Ptr)() = NULL;
 
 extern int (*FB_User_Ptr)( const int lv, const double TimeNew, const double TimeOld, const double dt,
-                           const int NPar, const int *ParSortID, real_par *ParAtt[PAR_NATT_TOTAL],
+                           const int NPar, const long *ParSortID, real_par *ParAtt[PAR_NATT_TOTAL],
                            real (*Fluid)[FB_NXT][FB_NXT][FB_NXT], const double EdgeL[], const double dh, bool CoarseFine[],
                            const int TID, RandomNumber_t *RNG );
